simplify for pr
