--- frontend/app/table/layout.tsx
@@ -1,35 +1,0 @@
-import '@/app/globals.css';
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-import { Sidebar } from '@/components/sidebar';
-import { ThemeProvider } from '@/components/theme-provider';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Domain Permutations',
-  description: 'Domain availability checker and permutation analyzer',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={inter.className} suppressHydrationWarning>
-        <ThemeProvider attribute="class" defaultTheme="dark" enableSystem={false}>
-          <div className="flex">
-            <div className="w-64">
-              <Sidebar />
-            </div>
-            <main className="flex-1 p-8">
-              {children}
-            </main>
-          </div>
-        </ThemeProvider>
-      </body>
-    </html>
-  );
-}
